+++ src/NetworkDeferred.h
@@ -43,12 +43,12 @@ public:
 	struct ReplyData
 	{
 		QByteArray data;
-		QList<QNetworkReply::RawHeaderPair> headers;
+		const QNetworkReply* qReply;
 
-		ReplyData() {}
-		ReplyData(QByteArray data, QList<QNetworkReply::RawHeaderPair> headers) : data(data), headers(headers) {}
+		ReplyData() : qReply(nullptr) {}
+		ReplyData(QByteArray data, const QNetworkReply* qReply) : data(data), qReply(qReply) {}
 
-		bool operator==(const ReplyData& other) const { return data == other.data && headers == other.headers; }
+		bool operator==(const ReplyData& other) const { return data == other.data && qReply == other.qReply; }
 	};
 
 	struct Progress
@@ -71,17 +71,18 @@ public:
 	{
 		QNetworkReply::NetworkError code;
 		QString message;
+		ReplyData replyData;
 
 		Error() : code(QNetworkReply::NoError) {}
-		Error(QNetworkReply::NetworkError code, const QString& message) : code(code), message(message) {}
+		Error(const ReplyData& replyData) : code(replyData.qReply->error()), message(replyData.qReply->errorString()), replyData(replyData) {}
 
 		bool operator==(const Error& other) const { return code == other.code && message == other.message; }
 	};
 
 	static Ptr create(QNetworkReply* reply);
 
-	ReplyData replyData() const { QMutexLocker locker(&m_lock); return ReplyData{m_buffer, m_reply->rawHeaderPairs()}; }
-	Error error() const { QMutexLocker locker(&m_lock); return Error{m_reply->error(), m_reply->errorString()}; }
+	ReplyData replyData() const { QMutexLocker locker(&m_lock); return ReplyData(m_buffer, m_reply); }
+	Error error() const { QMutexLocker locker(&m_lock); return Error(ReplyData(m_buffer, m_reply)); }
 
 signals:
 	void resolved(const QtPromise::NetworkDeferred::ReplyData& data) const;
@@ -95,6 +96,7 @@ private slots:
 	void replyFinished();
 	void replyDownloadProgress(qint64 bytesReceived, qint64 bytesTotal);
 	void replyUploadProgress(qint64 bytesSent, qint64 bytesTotal);
+	void replyDestroyed();
 
 private:
 	mutable QMutex m_lock;